3 tasks completes
Fetching Database data: do the TODO in files mockAPI/allArticles.js and src/Pages/Home.jsx to get the articles from the database to the react components.
Show the image, title, and summary in src/Components/ArticleCard/Article.jsx. for that you will have to do the TODO in Summary.jsx and TitleSection.jsx in the ArticleCard folder
Truncate article text in src/Components/ArticleCard/Summary.jsx to a max length. It is your choice to decide how the max length should be decided.